ProyectoFinal/requirements.txt
# Dependencias para Sistema de Canceladora Metro con InfluxDB

# Framework Web
Flask==3.0.0

# Hardware - Raspberry Pi
RPi.GPIO==0.7.0
gpiozero==2.0.1
RPLCD==1.3.1

# Comunicación Serial
pyserial==3.5

# InfluxDB
influxdb-client==1.36.0

# Utilidades
python-dotenv==1.0.0
requests==2.31.0
